% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TempleMetrics.R
\name{distreg}
\alias{distreg}
\title{title distreg}
\usage{
distreg(yvals, data, yname, xnames)
}
\arguments{
\item{yvals}{all the values of y to compute F(y|x)}

\item{data}{the dataset}

\item{yname}{the name of the y variables in the dataset}

\item{xnames}{the names of the x variables in the dataset}
}
\value{
DR object
}
\description{
the main function for running distribution regressions
}
\examples{
data(igm)
y0 <- median(igm$lcfincome)
distreg(y0, igm, "lcfincome", c("lfincome", "HEDUC"))

}
