% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TempleMetrics.R
\name{DR}
\alias{DR}
\title{DR}
\usage{
DR(yvals, glmlist)
}
\arguments{
\item{yvals}{the values of the y of F(y|x)}

\item{glmlist}{an estimated model for each y value for F(y|x)}
}
\description{
DR (distribution regression) objects
}
