% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TempleMetrics.R
\name{drs.inner}
\alias{drs.inner}
\title{drs.inner}
\usage{
drs.inner(y, data, yname, xnames, link = "logit")
}
\arguments{
\item{data}{the dataset}

\item{link}{which link function to use, it can be anything accepted
by glm (for example, logit, probit, or cloglog), the default is "logit"}
}
\value{
glm object
}
\description{
internal function that does the heavy lifting
 on estimating distribution regressions
}
\keyword{internal}
