% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TempleMetrics.R
\name{drs}
\alias{drs}
\title{drs}
\usage{
drs(yvals, data, yname, xnames)
}
\arguments{
\item{yvals}{all the values of y to compute F(y|x)}

\item{data}{the dataset}
}
\value{
list of glm objects for each value of y
}
\description{
run multiple distribution regressions
}
\keyword{internal}
