% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TempleMetrics.R
\name{Fycondx.llDRlist}
\alias{Fycondx.llDRlist}
\title{Fycondx.llDRlist}
\usage{
\method{Fycondx}{llDRlist}(object, yvals, xdf)
}
\arguments{
\item{object}{either a distribution regression or quantile regression object}

\item{yvals}{the values to compute the ecdf for}

\item{xdf}{a dataframe (can contain multiple rows) with x values}
}
\value{
a list of ecdfs for each row in xdf
}
\description{
take a particular value of y and predict F(y|x)
}
