% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TempleMetrics.R
\name{Fycondx.rqs}
\alias{Fycondx.rqs}
\title{Fycondx.rqs}
\usage{
\method{Fycondx}{rqs}(object, yvals, xdf)
}
\arguments{
\item{object}{a quantile regression object that has been estimated in a
first step}

\item{yvals}{the values to compute the ecdf for}

\item{xdf}{a dataframe (can contain multiple rows) with x values}
}
\value{
a list of conditional distributions
}
\description{
compute the conditional distribution of y conditional on x
 using quantile regression
}
