% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lldr.R
\name{k}
\alias{k}
\title{k}
\usage{
k(z, h = 1, type = "gaussian")
}
\arguments{
\item{z}{evaluate kernel at}

\item{h}{bandwidth}

\item{type}{either "gaussian" or "epanechnikov"}
}
\value{
k(z/h)
}
\description{
kernel function
}
\keyword{internal}
