% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lldr.R
\name{lldr.inner.y}
\alias{lldr.inner.y}
\title{lldr.inner.y}
\usage{
lldr.inner.y(y, xmain.seq, Y, XMain, XOther = NULL, h = NULL, method = "level")
}
\arguments{
\item{y}{a particular value of y to compute local linear distribution
regression for}

\item{xmain.seq}{a vector of x values to compute F(y|x) using local
linear distribution regression}

\item{Y}{a vector containing the data for the outcome}

\item{XMain}{a vector containing the data for the "main" x variable}

\item{XOther}{a matrix or data.frame containing the data for the "other"
x variables}

\item{h}{optional bandwidth}

\item{method}{"level" or "rank" determining whether method should
be used conditional on ytmin1 or the rank of ytmin1}
}
\value{
a list of llDR objects
}
\description{
calls lldr.inner.xvals for a particular value of y.  This is a
 dummy function that just reverses the arguments of lldr.inner.xvals in
 order to be able to call it for a vector of y values
}
\keyword{internal}
