% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TempleMetrics.R
\name{llDR}
\alias{llDR}
\title{llDR}
\usage{
llDR(y, t, thet)
}
\arguments{
\item{y}{the value of y for which F(y|t,x) was computed}

\item{t}{the value of t for which F(y|t,x) was computed}

\item{thet}{the local parameters from the local linear distribution
regression}
}
\description{
llDR (local linear distribution regression) object.
 It contains a value for y, a value for t, and a value for the parameters
}
