% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lldr.R
\name{lldrs.inner}
\alias{lldrs.inner}
\title{lldrs.inner}
\usage{
lldrs.inner(
  y.seq,
  xmain.seq,
  Y,
  XMain,
  XOther = NULL,
  h = NULL,
  method = "level",
  cl = 1
)
}
\arguments{
\item{y.seq}{a vector of y values to compute F(y|x) for using local linear
distribution regression}

\item{xmain.seq}{a vector of x values to compute F(y|x) using local
linear distribution regression}

\item{Y}{a vector containing the data for the outcome}

\item{XMain}{a vector containing the data for the "main" x variable}

\item{XOther}{a matrix or data.frame containing the data for the "other"
x variables}

\item{h}{optional bandwidth}

\item{method}{"level" or "rank" determining whether method should
be used conditional on ytmin1 or the rank of ytmin1}

\item{cl}{The number of clusters to use, default is 1}
}
\value{
a list of llDR objects that are indexed by y and t separately
}
\description{
internal function for running local linear distribution
 regression for a vector of y and x
}
