% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llscm.R
\name{llscm.inner}
\alias{llscm.inner}
\title{llscm.inner}
\usage{
llscm.inner(t, Y, T, Xmat, h)
}
\arguments{
\item{t}{conditional at a value T=t}

\item{Y}{outcome variable}

\item{T}{treatment variable}

\item{Xmat}{covariates}

\item{h}{bandwidth}
}
\value{
a 2*k (k being the dimension of X) vector of coefficients, the first
 k are the "levels", the second k are the derivatives with respect to each
 element of X.
}
\description{
internal function that does the heavy lifting
}
\keyword{internal}
