% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llscm.R
\name{llscm}
\alias{llscm}
\title{llscm}
\usage{
llscm(t, Y, T, Xmat, h)
}
\arguments{
\item{t}{conditional at a value T=t}

\item{Y}{outcome variable}

\item{T}{treatment variable}

\item{Xmat}{covariates}

\item{h}{bandwidth}
}
\value{
Coefficients
}
\description{
local linear estimator of smoothing coefficient model
}
