% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lldr.R
\name{wgr}
\alias{wgr}
\title{wgr}
\usage{
wgr(bet, y, xmain, xother = NULL, thisx, h)
}
\arguments{
\item{bet}{value of parameters}

\item{y}{vector of y values}

\item{thisx}{scalar value of x for which evaluating}

\item{h}{bandwidth}
}
\value{
weighted gradient
}
\description{
weighted gradient (wscore takes care of weights)
}
\keyword{internal}
