% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lldr.R
\name{wll}
\alias{wll}
\title{wll}
\usage{
wll(bet, y, xmain, xother = NULL, thisx, h)
}
\arguments{
\item{bet}{value of parameters}

\item{y}{vector of y values}

\item{thisx}{scalar value of x for which evaluating}

\item{h}{bandwidth}

\item{x}{vector of x values}
}
\value{
negative of log likelihood function
}
\description{
weighted log likelihood function for local logit model
 for scalar x only
}
\keyword{internal}
