% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TempleMetrics.R
\name{distreg}
\alias{distreg}
\title{distreg}
\usage{
distreg(formla, data, yvals, link = "logit")
}
\arguments{
\item{formla}{y ~ x}

\item{data}{the dataset}

\item{yvals}{all the values of y to compute F(y|x)}

\item{link}{which link function to use, it can be anything accepted
by glm (for example, logit, probit, or cloglog), the default is "logit"}
}
\value{
DR object
}
\description{
the main function for running distribution regressions
}
\examples{
data(igm)
y0 <- median(igm$lcfincome)
distreg(lcfincome ~ lfincome + HEDUC, igm, y0)

}
