% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TempleMetrics.R
\name{drs.inner}
\alias{drs.inner}
\title{drs.inner}
\usage{
drs.inner(y, data, yname, xnames)
}
\arguments{
\item{data}{the dataset}

\item{yname}{the name of the y variables in the dataset}

\item{xnames}{the names of the x variables in the dataset}
}
\value{
glm object
}
\description{
internal function that does the heavy lifting
 on estimating distribution regressions
}
\keyword{internal}
